+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Supported small thumbnail images in image gallery view for better navigation. (#495)
 - Supported displaying image filtered by category in image gallery view. (#487, #490)
+- Supported translating from traditional Chinese to simplified Chinese and vice versa. (#511)
 
 ### Changed